CB-555 refactor: add ProductCardItem components
공통적으로 사용되는 마크업을 상품 컴포넌트 분리
- 기존의 스와이퍼에서 사용되는 요소와 상품 상세페이지에서 사용
