+++ src/pages/Main/components/SwiperProductItem.tsx
@@ -1,8 +1,6 @@
-import { useNavigate } from 'react-router-dom';
-import ContentsContainer from '@/components/ContentsContainer';
 import { concatClasses } from '@/utils/libs/concatClasses';
 import { ProductPreviewType } from '@/@type/product';
-import { VerticalBox, VerticalCenterBox } from '../index.style';
+import ProductCardItem from '@/components/Product/ProductCardItem';
 
 interface SwiperProductItemProps {
   product: ProductPreviewType;
@@ -10,10 +8,6 @@ interface SwiperProductItemProps {
 }
 
 export default function SwiperProductItem({ product, isActive }: SwiperProductItemProps) {
-  const navigate = useNavigate();
-  const { name, brand, productId, thumbnail } = product;
-  const goProductDetailPage = () => navigate(`/products/${productId}`);
-
   return (
     <div
       className={concatClasses(
@@ -21,23 +15,7 @@ export default function SwiperProductItem({ product, isActive }: SwiperProductIt
         isActive ? 'scale-110' : 'scale-90',
       )}
     >
-      <ContentsContainer>
-        <VerticalBox className="flex-1 justify-between relative overflow-hidden text-ellipsis whitespace-nowrap">
-          <div className="min-h-[60%]">
-            <VerticalCenterBox onClick={goProductDetailPage}>
-              <img className="rounded-t w-full h-full" src={thumbnail} alt={name} />
-            </VerticalCenterBox>
-          </div>
-          <div className="p-1 flex flex-col flex-1 justify-between">
-            <VerticalBox className="overflow-hidden py-1 space-y-1" onClick={goProductDetailPage}>
-              <p className="text-[11px] leading-[11px] text-primary-dark font-semibold">{brand}</p>
-              <p className="text-[12px] leading-[12px] w-full whitespace-pre-line">
-                {`${name.slice(0, 28)}${name.length > 28 ? '...' : ''}`}
-              </p>
-            </VerticalBox>
-          </div>
-        </VerticalBox>
-      </ContentsContainer>
+      <ProductCardItem product={product} />
     </div>
   );
 }
